+++ examples/Data/DataB.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Flow\Examples\Data;
+
+readonly class DataB
+{
+    public function __construct(public int $id, public int $d, public int $e) {}
+}