+++ src/Flow/YFlow.php
@@ -7,8 +7,8 @@ namespace Flow\Flow;
 use Closure;
 use Flow\DriverInterface;
 use Flow\ExceptionInterface;
-use Flow\Ip;
 use Flow\IpStrategyInterface;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * @template T1
@@ -19,15 +19,15 @@ use Flow\IpStrategyInterface;
 class YFlow extends FlowDecorator
 {
     /**
-     * @param null|Closure(Ip<T1>, ExceptionInterface): void $errorJob
-     * @param null|IpStrategyInterface<T1>                   $ipStrategy
-     * @param null|DriverInterface<T1,T2>                    $driver
+     * @param null|Closure(ExceptionInterface): void $errorJob
+     * @param null|IpStrategyInterface<T1>           $ipStrategy
+     * @param null|DriverInterface<T1,T2>            $driver
      */
-    public function __construct(Closure $job, Closure $errorJob = null, IpStrategyInterface $ipStrategy = null, DriverInterface $driver = null)
+    public function __construct(Closure $job, ?Closure $errorJob = null, ?IpStrategyInterface $ipStrategy = null, ?EventDispatcherInterface $dispatcher = null, ?DriverInterface $driver = null)
     {
         $U = static fn (Closure $f) => $f($f);
         $Y = static fn (Closure $f) => $U(static fn (Closure $x) => $f(static fn ($y) => $U($x)($y)));
 
-        parent::__construct(new Flow($Y($job), $errorJob, $ipStrategy, $driver));
+        parent::__construct(new Flow($Y($job), $errorJob, $ipStrategy, $dispatcher, $driver));
     }
 }