+++ src/IpStrategyInterface.php
@@ -4,23 +4,9 @@ declare(strict_types=1);
 
 namespace Flow;
 
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+
 /**
  * @template T
  */
-interface IpStrategyInterface
-{
-    /**
-     * @param Ip<T> $ip
-     */
-    public function push(Ip $ip): void;
-
-    /**
-     * @return null|Ip<T>
-     */
-    public function pop(): ?Ip;
-
-    /**
-     * @param Ip<T> $ip
-     */
-    public function done(Ip $ip): void;
-}
+interface IpStrategyInterface extends EventSubscriberInterface {}