+++ tests/Driver/DriverTestCase.php
@@ -6,7 +6,12 @@ namespace Flow\Test\Driver;
 
 use Exception;
 use Flow\DriverInterface;
+use Flow\Event\PushEvent;
+use Flow\Exception\RuntimeException;
+use Flow\Ip;
+use Flow\IpStrategyEvent;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\EventDispatcher\EventDispatcher;
 
 /**
  * @template T1
@@ -17,68 +22,72 @@ abstract class DriverTestCase extends TestCase
     public function testAsync(): void
     {
         $driver = $this->createDriver();
-        $driver->async(static function () {
-        }, function (mixed $value) use ($driver) {
-            $driver->stop();
-            $this->assertNull($value);
-        })();
-        $driver->start();
+        $value = $driver->async(static function () {})();
+        self::assertNull($value);
+    }
+
+    /*public function testAsync(): void
+    {
+        $driver = $this->createDriver();
+        $stream = $driver->async(static function () {})();
+        $value = $driver->await($stream);
+        self::assertNull($value);
     }
 
     public function testAsyncReturn(): void
     {
         $driver = $this->createDriver();
-        $driver->async(static function () {
+        $stream = $driver->async(static function () {
             return 2;
-        }, function (mixed $value) use ($driver) {
-            $driver->stop();
-            $this->assertSame(2, $value);
         })();
-        $driver->start();
+        $value = $driver->await($stream);
+        self::assertSame(2, $value);
     }
 
     public function testAsyncError(): void
     {
         $driver = $this->createDriver();
-        $driver->async(static function () {
+        $stream = $driver->async(static function () {
             throw new Exception();
-        }, function (mixed $value) use ($driver) {
-            $driver->stop();
-            $this->assertInstanceOf(Exception::class, $value);
         })();
-        $driver->start();
+        $value = $driver->await($stream);
+        self::assertInstanceOf(RuntimeException::class, $value);
+        self::assertInstanceOf(Exception::class, $value->getPrevious());
     }
 
     public function testDelay(): void
     {
         $driver = $this->createDriver();
-        $driver->async(static function () use ($driver) {
+        $stream = $driver->async(static function () use ($driver) {
             $driver->delay(1 / 1000);
-        }, function (mixed $value) use ($driver) {
-            $driver->stop();
-            $this->assertNull($value);
         })();
-        $driver->start();
+        $value = $driver->await($stream);
+        self::assertNull($value);
     }
 
     public function testTick(): void
     {
-        self::assertTrue(true);
-
-        /*$i = 0;
+        $i = 0;
         $driver = $this->createDriver();
-        $driver->tick(1, function () use (&$i) {
+        $cancel = $driver->tick(1, static function () use (&$i) {
             $i++;
         });
-        $driver->async(function () use ($driver, &$i) {
-            $driver->delay(3 / 1000);
-            $driver->stop();
 
-            $this->assertGreaterThan(3, $i);
-        })();
+        $dispatcher = new EventDispatcher();
+        $dispatcher->dispatch(new PushEvent(new Ip()), IpStrategyEvent::PUSH);
 
-        $driver->start();*/
-    }
+        $stream = [
+            'ips' => 1,
+            'fnFlows' => [static function () use ($driver) {
+                $driver->delay(1 / 1000);
+            }],
+            'dispatchers' => [$dispatcher],
+        ];
+        $driver->await($stream);
+        $cancel();
+
+        self::assertGreaterThan(3, $i);
+    }*/
 
     /**
      * @return DriverInterface<T1,T2>