+++ tests/Flow/FlowTest.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace Flow\Test\Flow;
 
 use ArrayObject;
-use Closure;
+use Flow\Driver\AmpDriver;
 use Flow\DriverInterface;
+use Flow\ExceptionInterface;
 use Flow\Flow\Flow;
 use Flow\Ip;
 use Flow\IpStrategyInterface;
@@ -22,69 +23,75 @@ class FlowTest extends TestCase
     use FlowTrait;
 
     /**
-     * @dataProvider jobProvider
+     * @dataProvider provideJobCases
      *
      * @param DriverInterface<T1,T2>  $driver
      * @param IpStrategyInterface<T1> $ipStrategy
-     * @param array<Closure>          $jobs
+     * @param array<mixed>            $jobs
      */
     public function testJob(DriverInterface $driver, IpStrategyInterface $ipStrategy, array $jobs, int $resultNumber): void
     {
-        $ip = new Ip(new ArrayObject(['number' => 0]));
         $flow = array_reduce(
-            array_map(static fn ($job) => new Flow($job, static function () {}, $ipStrategy, $driver), $jobs),
+            array_map(static fn ($job) => new Flow(
+                $job,
+                static function (ExceptionInterface $exception) {
+                    self::assertSame(RuntimeException::class, $exception->getPrevious()::class);
+                },
+                $ipStrategy,
+                null,
+                $driver
+            ), $jobs),
             static fn ($flow, $flowIt) => $flow ? $flow->fn($flowIt) : $flowIt
         );
-        ($flow)($ip, static function (Ip $ip) use ($driver, $resultNumber) {
-            $driver->stop();
-
-            self::assertSame(ArrayObject::class, $ip->data::class);
-            self::assertSame($resultNumber, $ip->data['number']);
+        $flow->fn(static function (ArrayObject $data) use ($resultNumber) {
+            self::assertSame(ArrayObject::class, $data::class);
+            self::assertSame($resultNumber, $data['number']);
         });
+        $ip = new Ip(new ArrayObject(['number' => 0]));
+        ($flow)($ip);
 
-        $driver->start();
+        $flow->await();
     }
 
     /**
-     * @dataProvider jobProvider
+     * @dataProvider provideJobCases
      *
-     * @param DriverInterface<T1,T2|void> $driver
+     * @param DriverInterface<T1,T2>  $driver
+     * @param IpStrategyInterface<T1> $ipStrategy
+     * @param array<mixed>            $jobs
      */
-    public function testJobs(DriverInterface $driver): void
+    public function testTick(DriverInterface $driver, IpStrategyInterface $ipStrategy, array $jobs, int $resultNumber): void
     {
-        $ip1 = new Ip(new ArrayObject(['n1' => 3, 'n2' => 4]));
-        $ip2 = new Ip(new ArrayObject(['n1' => 2, 'n2' => 5]));
+        $cancel = $driver->tick(1, static function () use (&$flow) {
+            $ip = new Ip(new ArrayObject(['number' => 0]));
+            ($flow)($ip); // @phpstan-ignore-line
+        });
 
-        $jobs = [static function (object $data): void {
-            $data['n1'] *= 2;
-        }, static function (object $data): void {
-            $data['n2'] *= 4;
-        }];
-        $errorJobs = [static function () {
-        }, static function () {
-        }];
-        $flow = new Flow($jobs, $errorJobs, null, $driver);
+        $flow = array_reduce(
+            array_map(static fn ($job) => new Flow(
+                $job,
+                static function (ExceptionInterface $exception) use ($cancel) {
+                    self::assertSame(RuntimeException::class, $exception->getPrevious()::class);
+                    $cancel();
+                },
+                $ipStrategy,
+                null,
+                $driver
+            ), $jobs),
+            static fn ($flow, $flowIt) => $flow ? $flow->fn($flowIt) : $flowIt
+        );
+        $flow->fn(static function (ArrayObject $data) use ($resultNumber) {
+            self::assertSame(ArrayObject::class, $data::class);
+            self::assertSame($resultNumber, $data['number']);
 
-        $ips = new ArrayObject();
+            return $data;
+        });
 
-        $callback = function (Ip $ip) use ($driver, $ips, $ip1, $ip2) {
-            $ips->append($ip);
-            if ($ips->count() === 2) {
-                $driver->stop();
+        $flow->fn(static function () use ($cancel) {
+            $cancel();
+        });
 
-                $this->assertSame($ip1, $ips->offsetGet(0));
-                $this->assertSame($ip2, $ips->offsetGet(1));
-                self::assertSame(6, $ip1->data['n1']);
-                self::assertSame(16, $ip1->data['n2']);
-                self::assertSame(4, $ip2->data['n1']);
-                self::assertSame(20, $ip2->data['n2']);
-            }
-        };
-
-        ($flow)($ip1, $callback);
-        ($flow)($ip2, $callback);
-
-        $driver->start();
+        $flow->await();
     }
 
     /**
@@ -100,40 +107,37 @@ class FlowTest extends TestCase
         $flow = Flow::do($callable, [
             ...['driver' => $driver, 'ipStrategy' => $ipStrategy],
             ...($config ?? []),
-        ]);
+        ])->fn(static function (ArrayObject $data) use ($resultNumber) {
+            self::assertSame(ArrayObject::class, $data::class);
+            self::assertSame($resultNumber, $data['number']);
 
-        ($flow)($ip, static function (Ip $ip) use ($driver, $resultNumber) {
-            $driver->stop();
-
-            self::assertSame(ArrayObject::class, $ip->data::class);
-            self::assertSame($resultNumber, $ip->data['number']);
+            return $data;
         });
 
-        $driver->start();
+        ($flow)($ip);
+
+        $flow->await();
     }
 
     /**
      * @return array<array<mixed>>
      */
-    public static function jobProvider(): iterable
+    public static function provideJobCases(): iterable
     {
         $exception = new RuntimeException('job error');
 
         return self::matrix(static fn (DriverInterface $driver) => [
-            'oneJob' => [[static function (ArrayObject $data) {
+            'job' => [[static function (ArrayObject $data) {
                 $data['number'] = 5;
+
+                return $data;
             }], 5],
-            'oneAsyncJob' => [[static function (ArrayObject $data) use ($driver): void {
+            'asyncJob' => [[static function (ArrayObject $data) use ($driver) {
                 $driver->delay(1 / 1000);
                 $data['number'] = 5;
+
+                return $data;
             }], 5],
-            'twoAsyncJob' => [[static function (ArrayObject $data) use ($driver): void {
-                $driver->delay(1 / 1000);
-                $data['number'] += 5;
-            }, static function (ArrayObject $data) use ($driver): void {
-                $driver->delay(1 / 1000);
-                $data['number'] *= 2;
-            }], 10],
             'exceptionJob' => [[static function () use ($exception) {
                 throw $exception;
             }], 0],
@@ -146,12 +150,19 @@ class FlowTest extends TestCase
     public static function provideDoCases(): iterable
     {
         return self::matrix(static fn (DriverInterface $driver) => [
-            'simpleGenerator' => [static function () {
-                yield static function (ArrayObject $data) {
-                    $data['number'] = 5;
-                };
-                yield static function (ArrayObject $data) {
+            'simpleGenerator' => [static function () use ($driver) {
+                if ($driver::class !== AmpDriver::class) {
+                    yield static function (ArrayObject $data) {
+                        $data['number'] = 5;
+
+                        return $data;
+                    };
+                }
+                yield static function (ArrayObject $data) use ($driver) {
+                    $driver->delay(1 / 1000);
                     $data['number'] = 10;
+
+                    return $data;
                 };
             }, null, 10],
         ]);